refactor(cli): rename command handler modules
