+++ packages/opencode/src/cli/cmd/debug/agent.handler.ts
