+++ packages/opencode/src/cli/cmd/github.shared.ts
